+++ src/components/Header.jsx
@@ -175,26 +175,4 @@ const EmployeeList = () => {
   );
 };
 
-export default EmployeeList;
 
-    <div style={{ padding: '2rem', color: '#333' }}>
-      <h1>About Us</h1>
-      <p>This is the About page. Learn more about what we do!</p>
-    </div>
-  );import React from 'react';
-  
-  const About = () => {
-    return (
-      <div style={{ padding: '2rem', color: '#333' }}>
-        <h1>About Us</h1>
-        <p>This is the About page. Learn more about what we do!</p>
-      </div>
-    );
-  };
-  
-  export default About;
-  
-  
-};
-
-export default About;
